+++ src/main/java/no/novari/model/resource/AbstractCollectionResources.java
@@ -1,4 +1,4 @@
-package no.fint.model.resource;
+package no.novari.model.resource;
 
 import com.fasterxml.jackson.annotation.*;
 import com.fasterxml.jackson.core.type.TypeReference;
