+++ src/main/java/no/novari/model/resource/FintLinks.java
@@ -1,4 +1,4 @@
-package no.fint.model.resource;
+package no.novari.model.resource;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
