+++ src/main/java/no/novari/model/resource/Link.java
@@ -1,4 +1,4 @@
-package no.fint.model.resource;
+package no.novari.model.resource;
 
 import lombok.*;
