--- src/test/groovy/no/fint/model/testutils/Person.java
@@ -1,4 +1,0 @@
-package no.fint.model.testutils;
-
-public class Person {
-}
