+++ src/test/groovy/no/novari/model/resource/testutils/PersonResource.java
@@ -1,11 +1,11 @@
-package no.fint.model.resource.testutils;
+package no.novari.model.resource.testutils;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import no.fint.model.resource.FintLinks;
-import no.fint.model.resource.Link;
+import no.novari.model.resource.FintLinks;
+import no.novari.model.resource.Link;
 
 import java.util.List;
 import java.util.Map;
