+++ src/test/groovy/no/novari/model/testutils/Person.java
@@ -1,0 +1,4 @@
+package no.novari.model.testutils;
+
+public class Person {
+}
